backend/requirements.txt
# Django
Django>=5.0

# Django REST Framework para criar a API
djangorestframework>=3.12,<4.0

# Dependências adicionais para o Django Admin, autenticação, etc.
djangorestframework-simplejwt>=5.0,<6.0

# Dependência para suporte a CORS
django-cors-headers>=3.10,<4.0

# Para migrações do banco de dados e manipulação do ORM
#psycopg2>=2.9,<3.0  # Caso use PostgreSQL, se usar SQLite, não é necessário

# Biblioteca para testes de APIs
pytest-django>=4.0,<5.0

# Para processamento de dados e recomendações
scikit-learn>=1.0,<2.0    # Para algoritmos de aprendizado de máquina (caso utilize)
numpy>=1.21,<2.0          # Biblioteca para cálculos numéricos

# Para o gerenciamento de ambiente e variáveis de configuração
python-dotenv>=0.19,<1.0

# Dependência para verificar erros de lint e formatação
flake8>=4.0,<5.0

# Para lidar com testes em API REST (opcional)
requests>=2.25,<3.0

# Para gerar documentação da API
drf-yasg>=1.21.8

# Para dados de manipulação de datas e horários
pytz>=2022.0

# Caso deseje melhorar o desempenho de acesso ao banco de dados com cache
#django-redis>=5.0,<6.0

# Se você usar Celery para tarefas assíncronas
celery>=5.0,<6.0

# Bibliotecas para melhorar a segurança
#cryptography>=3.4,<4.0

# Dependência para manipulação de imagens
# Pillow==9.4.0

# Dependência para verificar CSRF
# django-csrf-cookie>=1.6,<2.0

# Ferramenta de monitoramento de erros
sentry-sdk>=2.0,<3.0

# Testes de cobertura com pytest
pytest-cov>=3.0,<4.0

# Para comunicação com Redis
# redis>=4.0,<5.0

# Caso use GraphQL
# graphene-django>=3.0,<4.0

# Para manipulação de arquivos Excel
openpyxl>=3.0,<4.0

# Para WebSockets (Django Channels)
channels>=4.0,<5.0

# Pandas para manipulação de dados
pandas>=1.5.3
